rename internal directory
